# docs/swagger.yaml
basePath: /api/v0
definitions:
...
  types.Movie:
    properties:
      adult:
        type: boolean
      backdrop_path:
        type: string
      genre_ids:
        items:
          type: integer
        type: array
      id:
        type: integer
      original_language:
        type: string
      original_title:
        type: string
      overview:
        type: string
      popularity:
        type: number
      poster_path:
        type: string
      release_date:
        type: string
      title:
        type: string
      video:
        type: boolean
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
...
  types.MovieResults:
    properties:
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/types.Movie'
        type: array
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  types.Show:
    properties:
      backdrop_path:
        type: string
      first_air_date:
        type: string
      genre_ids:
        items:
          type: integer
        type: array
      id:
        type: integer
      name:
        type: string
      origin_country:
        items:
          type: string
        type: array
      original_language:
        type: string
      original_name:
        type: string
      overview:
        type: string
      popularity:
        type: number
      poster_path:
        type: string
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
...
  types.ShowResults:
    properties:
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/types.Show'
        type: array
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  v0.MessageResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v0.SubtitleFilesResponse:
    properties:
      lang:
        type: string
      releasename:
        type: string
      subdata:
        type: string
      subencoding:
        type: string
      subformat:
        type: string
      subtitlename:
        type: string
      vttdata:
        type: string
    type: object
  v0.SubtitleFilesResultsResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/v0.SubtitleFilesResponse'
        type: array
      success:
        type: boolean
    type: object
...
  v0.TmdbMovieResultsResponse:
    properties:
      results:
        $ref: '#/definitions/types.MovieResults'
      success:
        type: boolean
    type: object
...
  v0.TmdbShowResultsResponse:
    properties:
      results:
        $ref: '#/definitions/types.ShowResults'
      success:
        type: boolean
    type: object
...
  v0.TorrentListResponse:
    properties:
      hash:
        type: string
      length:
        type: string
      name:
        type: string
    type: object
  v0.TorrentListResultsResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/v0.TorrentListResponse'
        type: array
      success:
        type: boolean
    type: object
  v0.TorrentStatsResponse:
    properties:
      downdata:
        type: string
      downpercent:
        type: string
      downspeed:
        type: string
      fulldata:
        type: string
      peers:
        type: string
      success:
        type: boolean
    type: object
host: localhost:9000
info:
  contact: {}
  title: Raven Torrent API
  version: 0.6.0
paths:
...
  /restart/downrate/{downrate}/uprate/{uprate}:
    get:
      parameters:
      - description: ' '
        in: path
        name: downrate
        required: true
        type: integer
      - description: ' '
        in: path
        name: uprate
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Restart torrent client with new bandwith limits
      tags:
      - Torrent client
  /startplayer/{base64path}/{base64args}:
    get:
      parameters:
      - description: ' '
        in: path
        name: base64path
        required: true
        type: string
      - description: ' '
        in: path
        name: base64args
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Launch media player application
      tags:
      - Media playback
  /stats/{hash}:
    get:
      parameters:
      - description: ' '
        in: path
        name: hash
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TorrentStatsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get torrent download stats
      tags:
      - Torrent client
...
  /subtitlesbyfile/{hash}/{base64path}/lang/{lang}:
    get:
      parameters:
      - description: ' '
        in: path
        name: hash
        required: true
        type: string
      - description: ' '
        in: path
        name: base64path
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.SubtitleFilesResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get subtitles by file hash
      tags:
      - Subtitle search
  /subtitlesbyimdb/{imdb}/lang/{lang}/season/{season}/episode/{episode}:
    get:
      parameters:
      - description: ' '
        in: path
        name: imdb
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      - description: ' '
        in: path
        name: season
        required: true
        type: integer
      - description: ' '
        in: path
        name: episode
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.SubtitleFilesResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get subtitles by IMDB id
      tags:
      - Subtitle search
  /subtitlesbytext/{text}/lang/{lang}/season/{season}/episode/{episode}:
    get:
      parameters:
      - description: ' '
        in: path
        name: text
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      - description: ' '
        in: path
        name: season
        required: true
        type: integer
      - description: ' '
        in: path
        name: episode
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.SubtitleFilesResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get subtitles by text
      tags:
      - Subtitle search
...
  /tmdbsearch/type/movie/lang/{lang}/page/{page}/text/{text}:
    get:
      parameters:
      - description: ' '
        in: path
        name: text
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      - description: ' '
        in: path
        name: page
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TmdbMovieResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Search movies
      tags:
      - Media search
  /tmdbsearch/type/tv/lang/{lang}/page/{page}/text/{text}:
    get:
      parameters:
      - description: ' '
        in: path
        name: text
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      - description: ' '
        in: path
        name: page
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TmdbShowResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Search shows
      tags:
      - Media search
  /torrents:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TorrentListResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get list of added torrents
      tags:
      - Torrent client
...
swagger: "2.0"